rest/requirements.txt
django
djangorestframework
pygments
psycopg2